movies controller unit tests

+++ src/controllers/movies.js
@@ -3,10 +3,6 @@ const Movies = require('../models/movies')
 const geocoder = require('../utils/geocoder')
 const asyncHandler = require('express-async-handler')
 
-const sendJSONresponse = (res, status, content) => {
-  res.status(status).json(content)
-}
-
 /**
  * @desc List movies
  * @route GET /movies
@@ -21,7 +17,7 @@ const moviesFindAll = asyncHandler(async (req, res) => {
     })
   } catch (e) {
     res.status(500).json({
-      error: 'Server error'
+      message: e.message
     })
   }
 })
@@ -31,15 +27,25 @@ const moviesFindAll = asyncHandler(async (req, res) => {
  * @route GET /movies/:movieid
  * @acces public */
 const moviesReadOne = asyncHandler(async (req, res) => {
+  if (!req.params.movieid) {
+    return res.status(400).json({
+      message: 'Movieid is required'
+    })
+  }
   try {
     const movie = await Movies.findById(req.params.movieid)
+    if (!movie) {
+      return res.status(404).json({
+        message: 'Movie not found'
+      })
+    }
     return res.status(200).json({
       success: true,
       data: movie
     })
   } catch (e) {
-    res.status(404).json({
-      error: 'movie not found'
+    res.status(400).json({
+      message: e.message
     })
   }
 })
@@ -50,30 +56,31 @@ const moviesReadOne = asyncHandler(async (req, res) => {
  * @acces private */
 const moviesCreate = asyncHandler(async (req, res) => {
   if (!req.body.location) {
-    sendJSONresponse(res, 400, { message: 'location is required' })
-  } else {
-    try {
-      const loc = await geocoder.geocode(req.body.location)
-      const parseLocation = {
-        type: 'Point',
-        coordinates: [loc[0].longitude, loc[0].latitude],
-        formattedLocation: loc[0].formattedAddress
-      }
-      const movie = await Movies.create({
-        title: req.body.title,
-        year: req.body.year,
-        genre: req.body.genre,
-        poster: req.body.poster,
-        rating: req.body.rating,
-        geoLocation: parseLocation
-      })
-      return res.status(201).json({
-        success: true,
-        data: movie
-      })
-    } catch (e) {
-      res.status(406).json(e.message)
+    return res.status(400).json({
+      message: 'Location is required'
+    })
+  }
+  try {
+    const loc = await geocoder.geocode(req.body.location)
+    const parseLocation = {
+      type: 'Point',
+      coordinates: [loc[0].longitude, loc[0].latitude],
+      formattedLocation: loc[0].formattedAddress
     }
+    const movie = await Movies.create({
+      title: req.body.title,
+      year: req.body.year,
+      genre: req.body.genre,
+      poster: req.body.poster,
+      rating: req.body.rating,
+      geoLocation: parseLocation
+    })
+    return res.status(201).json({
+      success: true,
+      data: movie
+    })
+  } catch (e) {
+    return res.status(406).json({ message: e.message })
   }
 })
 
@@ -82,42 +89,38 @@ const moviesCreate = asyncHandler(async (req, res) => {
  * @route PUT /movies/:moviesid
  * @acces private */
 const moviesUpdateOne = asyncHandler(async (req, res) => {
-  if (!req.body.location) {
-    sendJSONresponse(res, 400, { message: 'location is required' })
-  } else {
-    try {
-      const loc = await geocoder.geocode(req.body.location)
-      const parseLocation = {
-        type: 'Point',
-        coordinates: [loc[0].longitude, loc[0].latitude],
-        formattedLocation: loc[0].formattedAddress
-      }
-      Movies.findOneAndUpdate({ id: req.params.movieid }, { new: true }).exec(
-        (err, movie) => {
-          if (!movie) {
-            sendJSONresponse(res, 404, { message: 'movie not found' })
-          } else if (err) {
-            sendJSONresponse(res, 406, err._message)
-          }
-          movie.title = req.body.title
-          movie.year = req.body.year
-          movie.genre = req.body.genre
-          movie.poster = req.body.poster
-          movie.rating = req.body.rating
-          movie.geoLocation = parseLocation
-          // not updating reviews via this endpoint1
-          movie.save((err, movie) => {
-            if (err) {
-              sendJSONresponse(res, 406, err._message)
-            } else {
-              sendJSONresponse(res, 200, movie)
-            }
-          })
-        }
-      )
-    } catch (e) {
-      res.status(400).json(e.message)
+  if (!req.body.location || !req.params.movieid) {
+    return res.status(400).json({
+      message: 'Location and movieid are required'
+    })
+  }
+  try {
+    const loc = await geocoder.geocode(req.body.location)
+    const parseLocation = {
+      type: 'Point',
+      coordinates: [loc[0].longitude, loc[0].latitude],
+      formattedLocation: loc[0].formattedAddress
     }
+    const movie = await Movies.findOneAndUpdate({ id: req.params.movieid }, { new: true })
+    if (!movie) {
+      return res.status(404).json({
+        message: 'Movie not found'
+      })
+    }
+    movie.title = req.body.title
+    movie.year = req.body.year
+    movie.genre = req.body.genre
+    movie.poster = req.body.poster
+    movie.rating = req.body.rating
+    movie.geoLocation = parseLocation
+    // not updating reviews via this endpoint1
+    await movie.save(movie)
+    return res.status(200).json({
+      success: true,
+      data: movie
+    })
+  } catch (e) {
+    return res.status(406).json({ message: e.message })
   }
 })
 
@@ -126,14 +129,24 @@ const moviesUpdateOne = asyncHandler(async (req, res) => {
  * @route DELETE movies/:movieid
  * @acces private */
 const moviesDeleteOne = asyncHandler(async (req, res) => {
+  if (!req.params.movieid) {
+    return res.status(400).json({
+      message: 'Movieid is required'
+    })
+  }
   try {
-    await Movies.findByIdAndRemove(req.params.movieid)
+    const movie = await Movies.findByIdAndRemove(req.params.movieid)
+    if (!movie) {
+      return res.status(404).json({
+        message: 'Movie not found'
+      })
+    }
     return res.status(204).json({
       success: true,
       movieid: req.params.movieid
     })
   } catch (e) {
-    res.status(404).json(e.message)
+    return res.status(406).json({ message: e.message })
   }
 })
